serverless/freshbot/requirements.txt
requests==2.24.0
pyjwt==1.7.1
# boto3==1.9.130
# botocore==1.12.130
# certifi==2020.6.20
# chardet==3.0.4
# idna==2.10
# urllib3==1.25.10